+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## v0.6 - 2024-11-30
+
+### Fixed
+
+- Removed tcp optimization based on speedtest due to unstablity
+- Optimization and Set force apt to use IPv4 by default
+
 ## v0.5 - 2024-11-01
 
 ### Fixed